Consolidate TLD changes
